pyproject.toml
[tool.poetry]
name = "quokka-editor-back"
version = "0.1.0"
description = ""
authors = ["D0cza <d0.czajkowski@gmail.com>"]
readme = "README.md"
packages = [{include = "quokka_editor_back", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
gunicorn = "^20.1.0"
websockets = "^11.0.3"
aerich = "^0.7.1"
rich = "^13.3.5"
tortoise-orm = {version = "^0.19.3", extras = ["asyncpg"]}
starlette-admin = "^0.9.0"
pytest = "^7.3.2"
python-multipart = "^0.0.6"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
pydantic = {version = "^1.10.9", extras = ["email"]}
python3-crdt = "^1.0.3"
httpx = "^0.24.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.2"
isort = "^5.12.0"
pytest = "^7.3.1"
flake8 = "^6.0.0"
ipdb = "^0.13.13"
pre-commit = "^3.2.2"
mypy = "^1.4.0"
pytest-asyncio = "^0.21.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
env = [
    "DEBUG = false"
]

[tool.aerich]
tortoise_orm = "quokka_editor_back.settings.TORTOISE_ORM"
location = "./src/quokka_editor_back/models/migrations"
src_folder = "./src/quokka_editor_back/."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
